+++ RELEASE_NOTES.md
@@ -11,3 +11,4 @@
 ## Bug Fixes
 - Fill missing values in aggregated component columns.
 - Update aggregate metrics to have only positive values in grid consumption.
+- Update grid feed-in metrics to first include negative values from grid column and simplify the grid feed-in logic.